Move skc_runtime to packages/core/ext
