+++ Cargo.toml
@@ -7,7 +7,7 @@ default-run = "shiika"
 
 [workspace]
 members = [
-  "lib/*",
+  "lib/*", "packages/core/ext",
 ]
 
 [features]
