--- lib/skc_runtime/README.md
@@ -1,3 +1,0 @@
-# skc_runtime
-
--> doc/new_runtime.md
