--- lib/skc_runtime/src/core_class/object.rs
@@ -1,5 +1,0 @@
-use shiika_ffi::core_class::SkObject;
-use shiika_ffi_macro::shiika_method;
-
-#[shiika_method("Object#initialize")]
-pub extern "C" fn object_initialize(_receiver: SkObject) {}
