+++ packages/core/exports.json5
