+++ packages/core/ext/src/core_class.rs
