+++ packages/core/ext/src/core_class/int.rs
