+++ packages/core/index.sk
@@ -1,0 +1,1 @@
+# TODO: load builtin
